requirements.txt

torch>=2.1.0
torchvision>=0.16.0
einops>=0.7.0
timm>=0.9.12
datasets>=2.15.0
pillow>=10.0.0
numpy<2.0.0
lpips>=0.1.4
torchmetrics>=1.2.0
mlflow>=2.8.0
tqdm>=4.66.0
xformers>=0.0.22